MND 613
My journey using Datawrapper didn’t come without its trials and tribulations. I struggled with getting this assignment off the ground, as I was really set on pulling my own data instead of using the datasets already in Datawrapper; however, for me, that proved to be more difficult than anticipated.
In light of the start of the Chinese Lunar New Year, and after getting some inspiration on Google Trends, I decided to try to pull my own data and attempt to track the U.S. search interest in the top 5 luckiest/most popular Chinese zodiac animals: Dragon, Snake, Pig, Rat, and Tiger.
Closely following the tutorial titled, “How to get data from Google Trends for charts or maps,” I ran a search in Google Trends for those animals, using the search term, “Chinese zodiac animal dragon,” or “Chinese zodiac animal snake,” etc.”
I fluctuated between the date range I wanted to capture, starting with a default Google start date of Jan 2004 through the present, or even gathering search history from the last five to 10 years. I realized capturing the data with those parameters was tricky, as I couldn’t quite manipulate the graph or the data to reflect what I wanted. I tinkered with the data, per the tutorial, to “clean” up the spreadsheet a bit so it would have more success formatting (it did not). 
After a few unsuccessful attempts, and realizing I might be off-base, I bailed on that idea and pursued attempt number two: uploading data from a verified source. In this case, I found Data USA, and focused on my hometown and current city of Schenectady, NY. Here, there were quite a few datasets from which to pull, and I was drawn to one showcasing poverty by age and gender. Specifically, this data revealed that Females 25 - 34 are the largest demographic living in poverty in the city. This made me very sad, being a female and growing up in this town; there’s no denying its fabric and citizens have changed dramatically in the last 50 years, and while change is fantastic, it’s sad to see a new generation of citizens — women — without the support or opportunities they likely deserve.
Using data from the Census Bureau ACS 5-Year Estimate, the data shows, “19.4% of the population for whom poverty status is determined in Schenectady, NY (12.1k out of 62.3k people) live below the poverty line, a number that is higher than the national average of 12.3%. The largest demographic living in poverty are Females 25 - 34, followed by Females 18 - 24 and then Females 35 - 44” (Data USA, n.d.).
To provide a bit of context, the American Community Survey (ACS) is conducted by the U.S. Census and sent to a portion of the population every year (Data USA, n.d.). Additionally, the Census Bureau uses a set of money income thresholds that vary by family size and composition to determine who classifies as impoverished; if a family's total income is less than the family's threshold than that family and every individual in it is considered to be living in poverty (Data USA, n.d.).
I wanted to download the data from that particular example, and recreate a similar data visualization, but at this time, I am struggling with cleaning the data to create a decent visual. I was trying to illustrate the total city population (62.3k) with the population per gender/age group, or the percent of the population, and reflect that against the total population, for both men and women, but I was unsuccessful so far.The visual I’m uploading is the only one that closely resembled a chart. I’d like to continue to dig into this assignment/tool so I can better understand how it works, as I am feeling too confused at the moment. 

<iframe title="Poverty by Age and Gender | Schenectady, NY" aria-label="Stacked Bars" id="datawrapper-chart-tB2qU" src="https://datawrapper.dwcdn.net/tB2qU/1/" scrolling="no" frameborder="0" style="width: 0; min-width: 100% !important; border: none;" height="763"></iframe><script type="text/javascript">!function(){"use strict";window.addEventListener("message",(function(e){if(void 0!==e.data["datawrapper-height"]){var t=document.querySelectorAll("iframe");for(var a in e.data["datawrapper-height"])for(var r=0;r<t.length;r++){if(t[r].contentWindow===e.source)t[r].style.height=e.data["datawrapper-height"][a]+"px"}}}))}();
</script>